+++ app/reducers/types.ts
@@ -1,0 +1,11 @@
+import { Dispatch as ReduxDispatch, Store as ReduxStore, Action } from 'redux';
+
+export type counterStateType = {
+  counter: number;
+};
+
+export type GetState = () => counterStateType;
+
+export type Dispatch = ReduxDispatch<Action<string>>;
+
+export type Store = ReduxStore<counterStateType, Action<string>>;