+++ configs/webpack.config.main.prod.babel.js
@@ -9,17 +9,19 @@ import TerserPlugin from 'terser-webpack-plugin';
 import { BundleAnalyzerPlugin } from 'webpack-bundle-analyzer';
 import baseConfig from './webpack.config.base';
 import CheckNodeEnv from '../internals/scripts/CheckNodeEnv';
+import DeleteSourceMaps from '../internals/scripts/DeleteSourceMaps';
 
 CheckNodeEnv('production');
+DeleteSourceMaps();
 
 export default merge.smart(baseConfig, {
-  devtool: 'source-map',
+  devtool: process.env.DEBUG_PROD === 'true' ? 'source-map' : 'none',
 
   mode: 'production',
 
   target: 'electron-main',
 
-  entry: './app/main.dev',
+  entry: './app/main.dev.ts',
 
   output: {
     path: path.join(__dirname, '..'),
@@ -57,7 +59,8 @@ export default merge.smart(baseConfig, {
     new webpack.EnvironmentPlugin({
       NODE_ENV: 'production',
       DEBUG_PROD: false,
-      START_MINIMIZED: false
+      START_MINIMIZED: false,
+      E2E_BUILD: false
     })
   ],